refactor: split root command
